+++ src/myrrh/extensions/mregistry.py
@@ -3,12 +3,12 @@ import urllib.request
 import urllib.error
 import abc
 
-from ..protocol import StdExtSession
-from ..misc import URI, singleton
-from ..errors import ReadOnlyPath, InvalidPath
-from ..interfaces import ExtSessionT, IMyrrhExt, IExtSession, uri_rd
+from myrrh.exts.protocol import MyrrhExtBase, StdExtSession
+from myrrh.exts.misc import URI, singleton
+from myrrh.exts.errors import ReadOnlyPath, InvalidPath
+from myrrh.exts.interfaces import ExtSessionT, IMyrrhExt, IExtSession, uri_rd
 
-from .registry import Registry
+from myrrh.exts.registry import Registry
 
 
 class IExtRegistrySession(IExtSession):
@@ -29,25 +29,21 @@ class ExtRegistrySession(StdExtSession, IExtRegistrySession, typing.Generic[ExtS
         super().__init__(interface=IExtRegistrySession)
 
     def findall(self, prefix: str | None = None):
-        return Registry().findall(prefix=prefix)
+        return Registry().findall(uri=prefix)
 
     def loaded(self):
         return list(Registry().loaded)
 
 
-class ExtRegistry(IMyrrhExt[ExtRegistrySession]):
-
-    _path = "."
+class ExtRegistry(MyrrhExtBase):
 
     def open(self, uri: str, *, req: urllib.request.Request | None = None) -> ExtRegistrySession:
 
-        if str(URI(uri).path) != self._path:
+        if URI(uri).path != self._path:
             raise InvalidPath(URI(uri).path)
 
         return ExtRegistrySession()
 
-    def basepath(self, path: str):
-        self._path = path
 
     def extend(self, path: str, obj: typing.Any):
         raise ReadOnlyPath(self._path)