+++ src/myrrh/exts/client.py
@@ -8,12 +8,26 @@ import urllib.error
 from .misc import URI
 from .interfaces import ExtSessionT, IMyrrhExt
 
-
+class _Session:
+    def __init__(self, session):
+        self.session = session
+        self.closed = False
+        
+    def __del__(self):
+        self.session.close()
+        
+    def close(self):
+        if not self.closed:
+            self.session.close()
+        
+    def __getattr__(self, name: str):
+            return getattr(self.session, name)
+        
 class ExtClient(typing.Generic[ExtSessionT]):
 
     def __init__(self, serv: IMyrrhExt):
         self.serv = serv
-        self._uri = ""
+        self._uri = ''
 
     def seturi(self, uri):
         self._uri = uri
@@ -48,3 +62,9 @@ class ExtClient(typing.Generic[ExtSessionT]):
 
         with self.open(uri_.uri, req=req) as session:
             return session.request(uri_.split.query, data)
+        
+    def extend(self, path:str, obj: typing.Any):
+        if not path.startswith('/'):
+            path = '/'.join((self.serv.basepath(), path))
+            
+        self.serv.extend(path, obj)