+++ src/myrrh/exts/errors.py
@@ -11,6 +11,9 @@ class PathError(urllib.error.URLError): ...
 class URIOpenError(urllib.error.URLError): ...
 
 
+class InvalidRequest(urllib.error.URLError): ...
+
+
 class ExtendTypeError(ExtError):
     def __init__(self, type: typing.Any, path: str | None = None) -> None:
         super().__init__(f"invalid extension type {type.__class__}", path)