+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.2.0 - 2024-06-03
+
+- move sources under myrrh-core namespaces
+- updated APIs
+- fixed loading with sub-extension path
+- rewrote StdExtSession code
+- added a new api decorator to customize the name of "data" argument
+- added extend method to service client
+- findall parameter is uri instead of group name
+- auto load subpath
+
 ## 0.1.0 - 2024-06-03
 
 Initial version